Lower the memory consumption
A pako object contains a 64k buffer. We create a `FlateWorker` for each zip
entry, meaning a zip file with a lot of entries would take **a lot** of memory.

Lazy-loading the pako object isn't the best solution but it's the quickest.
The best solution is to lazy-load the worker list.

Mitigate the issue #446.